Leave profile decoding as an exercise for the user
Okta has a complex profile system where users can add their own attributes. This means we can't anticipate the names of the fields that might be present in the profile. In order to avoid confusion about that, this change removes parsing of the UserProfile and leaves it as an exercise for the user.